refactor complexity out of user-data & into ansible roles

+++ ansible/roles/aws_docker_config/files/scratch-docker.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
-# Assemble an EC2 instance's instance storage volumes into a RAID0 array and mount it to
-#     /mnt/scratch
-# when deployed as user-data, the script's log goes to /var/log/cloud-init-output.log
+# To run on first boot of an EC2 instance with NVMe instance storage volumes:
+# 1) Assembles them into a RAID0 array, formats with XFS, and mounts to /mnt/scratch
+# 2) Replaces /var/lib/docker with a symlink to /mnt/scratch/docker so that docker images and
+#    container file systems use this high-performance scratch space. (restarts docker)
+# The configuration persists through reboots (but not instance stop).
+# logs go to /var/log/cloud-init-output.log
 # refs:
 # https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ssd-instance-store.html
 # https://github.com/kislyuk/aegea/blob/master/aegea/rootfs.skel/usr/bin/aegea-format-ephemeral-storage
@@ -9,13 +12,10 @@
 set -euxo pipefail
 shopt -s nullglob
 
-if grep /dev/md0 <(df) ; then
-    exit 0
-fi
 
 devices=(/dev/xvd[b-m] /dev/disk/by-id/nvme-Amazon_EC2_NVMe_Instance_Storage_AWS?????????????????)
 num_devices="${#devices[@]}"
-if (( num_devices > 0 )); then
+if (( num_devices > 0 )) && ! grep /dev/md0 <(df); then
     mdadm --create /dev/md0 --force --auto=yes --level=0 --chunk=256 --raid-devices=${num_devices} ${devices[@]}
     mkfs.xfs -f /dev/md0
     mkdir -p /mnt/scratch
@@ -26,7 +26,7 @@ fi
 mkdir -p /mnt/scratch/tmp
 chown 1337:1337 /mnt/scratch /mnt/scratch/tmp
 
-# Move docker storage onto /mnt/scratch
+
 systemctl stop docker || true
 if [ -d /var/lib/docker ] && [ ! -L /var/lib/docker ]; then
     mv /var/lib/docker /mnt/scratch
